Improved logs

+++ micronaut/finerio-pfm-api/src/main/groovy/mx/finerio/pfm/api/dtos/AccountDto.groovy
@@ -12,7 +12,7 @@ class AccountDto  extends ResourceDto{
     String nature
     String name
     String number
-    float balance
+    BigDecimal balance
 
     AccountDto() {}
